combined_ligament_solver/requirements.txt
# Data processing and Excel handling
pandas>=1.5.0
openpyxl>=3.0.0
xlrd>=2.0.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Symbolic mathematics
sympy>=1.9.0

# MCMC sampling
emcee>=3.1.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration
PyYAML>=6.0

# Testing
pytest>=7.0.0
